isolate test setup func
